git init
website already mostly done... started with Monologue, a rails gem that adds a blog. had a LOT of issues integrating with static pages because I don't want to use the blog as the root (home page). this made dynamically serving up content much more difficult, as there is a root for the portfolio and a root for the blog (due to the nature of the monologue gem).